+++ practice/bank_system/src/main/java/bank/interfaces/IBankAccountFactory.java
@@ -3,5 +3,5 @@ package bank.interfaces;
 import bank.domains.BankAccount;
 
 public interface IBankAccountFactory {
-    BankAccount create(int id, String name, int balance);
+    BankAccount create(int id, String name, double balance);
 }
