+++ practice/bank_system/src/main/java/bank/interfaces/IBankAccountStorage.java
@@ -5,7 +5,7 @@ import bank.domains.BankAccount;
 import java.util.List;
 
 public interface IBankAccountStorage {
-    void addAccount(IBankAccountFactory bankAccountFactory, String name, int balance);
+    void addAccount(IBankAccountFactory bankAccountFactory, String name, double balance);
 
     void setAccounts(List<BankAccount> bankAccounts);
